+++ CHANGELOG.md
@@ -40,6 +40,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### State Breaking
+
+### Bug Fixes
+* [#6190](https://github.com/osmosis-labs/osmosis/pull/6190) v19 upgrade handler superfluid fix
+
+### Misc Improvements
+
+### Minor improvements & Bug Fixes
+
+### Security
+
 ## v18.0.0
 
 ### Misc Improvements